+++ packages/terra-embedded-content-consumer/tests/jest/EmbeddedContentConsumer.test.jsx
@@ -26,12 +26,19 @@ it('should render the embedded content consumer with custom class names', () => 
   expect(wrapper).toMatchSnapshot();
 });
 
-it('should render the embedded content consumer with `srcdoc` option', () => {
-  const frameOptions = { iframeAttrs: { title: 'inline html content', width: '100%', height: '100px' } };
+it('should render the embedded content consumer with options attributes', () => {
+  const inlineHtml = '<p><b>Inline HTML Content</b></p><p>This is an inline html content that can be used to render the content into the frame.</p>';
+  const frameOptions = {
+    iframeAttrs: {
+      title: 'inline html content',
+      width: '100%',
+      height: '100px',
+      srcdoc: inlineHtml,
+    },
+  };
   const embeddedContentConsumer = (
     <div>
       <EmbeddedContentConsumer
-        srcdoc="<p><b>Inline HTML Content</b></p><p>This is an inline html content that can be used to render the content into the frame.</p>"
         options={frameOptions}
       />
     </div>