+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-* Added
-  * Added `srcdoc` prop to allow inline html content to render inside the iframe.
+* Updated
+  * Updated documentation for `options` prop to include `srcdoc` iframe atttribute for inline html content in iframe.
 
 ## 3.37.0 - (July 31, 2023)